# projects/agent-kanban/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Kanban API
  description: |
    多项目 AI Agent 统一看板的 REST API。

    核心资源链：Workspace → Project → Task → Assignment → Agent

    **契约说明**：
    - Task 的创建/更新/关闭以 GitHub Issues 为准，本 API 提供只读聚合查询。
    - Assignment 的创建和管理通过本 API。
    - 此文件是 **canonical** API 契约，Go 和 Node.js 两个后端都基于此实现。
    - 所有时间字段为 ISO 8601 格式 (e.g. "2026-08-03T07:22:00Z")。
  version: "1.1.0"
  contact:
    name: Agent Kanban Team
servers:
  - url: http://localhost:8080/api/v1
    description: Go 后端 (本地开发)
  - url: http://localhost:4001/api/v1
    description: Node.js 后端 (本地开发)
  - url: http://localhost:4002/api/v1
    description: Next.js 嵌入式后端 (本地开发)
security:
  - ApiKeyAuth: []
  - {}
...
paths:
...
  /workspaces/{workspaceId}/projects:
    get:
      tags: [Projects]
      summary: 获取工作空间下的所有项目
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: 项目列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
              examples:
                default:
                  value:
                    - id: "proj-001"
                      workspaceId: "ws-001"
                      name: "前端项目"
                      repoOwner: "my-org"
                      repoName: "frontend-app"
                      syncStatus: "idle"
                      lastSyncedAt: "2026-08-03T06:00:00Z"
                      createdAt: "2026-08-01T00:00:00Z"
    post:
      tags: [Projects]
      summary: 添加项目（关联 GitHub Repo）
      operationId: addProject
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
            examples:
              default:
                value:
                  name: "前端项目"
                  repoOwner: "my-org"
                  repoName: "frontend-app"
      responses:
        "201":
          description: 添加成功，返回 webhook 配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithWebhook"
              examples:
                default:
                  value:
                    id: "proj-001"
                    workspaceId: "ws-001"
                    name: "前端项目"
                    repoOwner: "my-org"
                    repoName: "frontend-app"
                    syncStatus: "idle"
                    lastSyncedAt: null
                    createdAt: "2026-08-03T07:22:00Z"
                    webhookUrl: "http://localhost:8080/api/v1/webhooks/github"
                    webhookSecret: "whsec_abc123xyz"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: 获取项目详情
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: 移除项目
      operationId: removeProject
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: 移除成功
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/sync:
    post:
      tags: [Projects]
      summary: 手动触发项目同步
      operationId: triggerSync
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "202":
          description: 同步已触发
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"
              examples:
                default:
                  value:
                    projectId: "proj-001"
                    status: "queued"
                    message: "Full sync started"
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════
  # Tasks (read-only aggregation)
  # ═══════════════════════════════════════════════
  /workspaces/{workspaceId}/tasks:
    get:
      tags: [Tasks]
      summary: 获取工作空间的所有任务（跨项目聚合）
      operationId: listTasks
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: projectId
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: label
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: 任务列表（聚合视图）
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskWithAssignment"
              examples:
                default:
                  value:
                    total: 2
                    items:
                      - id: "task-001"
                        projectId: "proj-001"
                        issueNumber: 42
                        title: "实现用户登录页面"
                        body: "需要支持邮箱+密码登录"
                        labels: ["frontend", "ui"]
                        status: "backlog"
                        githubAssignee: null
                        githubUrl: "https://github.com/my-org/frontend-app/issues/42"
                        projectName: "前端项目"
                        repoOwner: "my-org"
                        repoName: "frontend-app"
                        currentAssignment:
                          id: "asg-001"
                          taskId: "task-001"
                          agentId: "agent-001"
                          status: "pending"
                          agent:
                            name: "前端开发专家"
                            modelName: "kimi-long-v1"
                        createdAt: "2026-08-02T10:00:00Z"
                        updatedAt: "2026-08-02T10:00:00Z"
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: 获取单个任务详情
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Tasks]
      summary: 更新任务状态（看板拖拽）
      description: |
        按 4 列白名单 + WIP 上限校验。成功后写回 GitHub `status:*` label（无 token 时跳过写回）。
        非法流转或 WIP 超限返回 409。
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
            examples:
              moveToReview:
                value:
                  status: in_review
      responses:
        "200":
          description: 更新后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 非法流转或 WIP 上限
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ═══════════════════════════════════════════════
  # Agents
  # ═══════════════════════════════════════════════
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 可选。Zero-Auth 模式下不需要。生产环境启用时在 header 传 X-API-Key。
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                code: "NOT_FOUND"
                message: "Workspace not found"
...
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            duplicate_project:
              value:
                code: "CONFLICT"
                message: "Project with repo my-org/frontend-app already exists"
            duplicate_assignment:
              value:
                code: "CONFLICT"
                message: "Task already has a pending or running assignment"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [NOT_FOUND, VALIDATION_ERROR, CONFLICT, UNAUTHORIZED, NO_MATCHING_AGENT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
    TaskStatus:
      type: string
      enum: [backlog, in_progress, in_review, done]
      description: "看板 4 列：待规划 / 进行中 / 待审核 / 已完成"
    AssignmentStatus:
      type: string
      enum: [pending, running, done, failed]
...
    Project:
      type: object
      required: [id, workspaceId, name, repoOwner, repoName, syncStatus, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        repoOwner:
          type: string
        repoName:
          type: string
        webhookSecret:
          type: string
          nullable: true
          description: 仅在创建时返回一次
        syncStatus:
          type: string
          enum: [idle, syncing, error]
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateProjectInput:
      type: object
      required: [name, repoOwner, repoName]
      properties:
        name:
          type: string
        repoOwner:
          type: string
        repoName:
          type: string
    ProjectWithWebhook:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            webhookUrl:
              type: string
            webhookSecret:
              type: string
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            taskCount:
              type: integer
    SyncStatus:
      type: object
      properties:
        projectId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        message:
          type: string
    Task:
      type: object
      required: [id, projectId, issueNumber, title, status, labels]
      properties:
        id:
          type: string
        projectId:
          type: string
        issueNumber:
          type: integer
        title:
          type: string
        body:
          type: string
        labels:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        githubAssignee:
          type: string
          nullable: true
        githubUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskWithAssignment:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            projectName:
              type: string
            repoOwner:
              type: string
            repoName:
              type: string
            currentAssignment:
              allOf:
                - $ref: "#/components/schemas/Assignment"
              nullable: true
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/TaskWithAssignment"
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/Assignment"
            project:
              $ref: "#/components/schemas/Project"
...
    UpdateTaskInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/TaskStatus"
...
    Assignment:
      type: object
      required: [id, taskId, agentId, status, retryCount, assignedAt]
      properties:
        id:
          type: string
        taskId:
          type: string
        agentId:
          type: string
        status:
          $ref: "#/components/schemas/AssignmentStatus"
        modelUsed:
          type: string
          nullable: true
        resultSummary:
          type: string
          nullable: true
        instruction:
          type: string
          nullable: true
        githubCommentId:
          type: integer
          nullable: true
        retryCount:
          type: integer
        assignedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        agent:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            role:
              type: string
            modelName:
              type: string
...